When copying, do not preserve timestamps.
As currently written, this copies across a compiled file if the source
hasn't changed since last time, but that file will not be used because
the FixMyStreet code checks timestamps and only uses compiled files if
newer than source files.

I think originally, the idea was to copy across the old compiled file,
then compare against the new source, but that won't work as the source
file will always be newer as it is a new checkout.

If we don't preserve timestamps, then if the source is the same as the
last deploy, the compiled will be the same and it can be copied across
as a new file.