Change load-mm script to accept list of files to load
Having the DATADIR environment variable restricts us to one directory of
gz files. We want to be able to load the files for multiple councils, so
this change changes things so we take a list of .gz files to load as
command line arguments and then iterate over those, rather than
iterating over the contents of a directory.