[Hackney] Add script to parse categories spreadsheet
During development, Hackney stored their list of FMS categories in a
Google spreadsheet which was continually iterated and improved. This
script parsed that sheet and output a JSON blob suitable for importing
into FixMyStreet using the `import_categories` script.

This script can also output a JavaScript file suitable for pasting into
the Hackney cobrand's `assets.js` file for all the WFS asset layers.

This script won't need to be run again and is included here mainly for
posterity and in case it's a useful basis for something similar in the
future.